+++ Changelog.md
@@ -22,6 +22,7 @@ OpenCore Changelog
 - Added `ResetTrafficClass` to reset TCSEL to T0 on legacy HDA
 - Fixed default boot entry selection without timeout for builtin picker
 - Added ocpasswordgen utility to generate OpenCore password data
+- Added `ActivateHpetSupport` quirk to activate HPET support
 
 #### v0.6.6
 - Added keyboard and pointer entry scroll support in OpenCanopy